[OBMIN-337] Extracted Child requests from User to an independent flow
