+++ src/main/java/space/obminyashka/items_exchange/dao/AdvertisementRepository.java
@@ -51,14 +51,14 @@ public interface AdvertisementRepository extends JpaRepository<Advertisement, UU
     Collection<Advertisement> findAllByUserUsername(String username);
 
     @Query("SELECT a from Advertisement a where " +
-            "(:id is null or a.id != :id) and " +
+            "(:id is null or a.id <> :id) and " +
             "(:subcategoryId is null or a.subcategory.id = :subcategoryId)")
     Page<Advertisement> findAllByIdNotAndSubcategoryId(@Param("id") UUID id,
                                                        @Param("subcategoryId") Long subcategoryId,
                                                        Pageable pageable);
 
     @Query("SELECT count(a) from Advertisement a where " +
-            "(:id is null or a.id != :id) and " +
+            "(:id is null or a.id <> :id) and " +
             "(:subcategoryId is null or a.subcategory.id = :subcategoryId)")
     Long countByIdNotAndSubcategoryId(@Param("id") UUID id, @Param("subcategoryId") Long subcategoryId);
 }
