+++ src/test/java/space/obminyashka/items_exchange/controller/AdvertisementControllerIntegrationTest.java
@@ -1,24 +1,30 @@
 package space.obminyashka.items_exchange.controller;
 
+import org.json.JSONObject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import space.obminyashka.items_exchange.BasicControllerTest;
 import space.obminyashka.items_exchange.dto.AdvertisementModificationDto;
 import space.obminyashka.items_exchange.util.AdvertisementDtoCreatingUtil;
-import space.obminyashka.items_exchange.util.ResponseMessagesHandler;
 import space.obminyashka.items_exchange.util.MessageSourceUtil;
+import space.obminyashka.items_exchange.util.ResponseMessagesHandler;
 
+import java.nio.charset.StandardCharsets;
 import java.util.Objects;
+import java.util.UUID;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static space.obminyashka.items_exchange.api.ApiKey.*;
 import static space.obminyashka.items_exchange.util.MessageSourceUtil.getMessageSource;
@@ -84,5 +90,33 @@ class AdvertisementControllerIntegrationTest extends BasicControllerTest {
         return MessageSourceUtil.getMessageSource(ResponseMessagesHandler.ValidationMessage.INVALID_MAX_SIZE)
                 .replace("${validatedValue}", dtoFieldValue)
                 .replace("{max}", maxValidValue);
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"notExisted", ""})
+    @WithMockUser("test")
+    void testDefaultLocalizationAccordingToLanguageHeader(String languageHeader) throws Exception {
+        MvcResult mvcResult = mockMvc.perform(delete(ADV_ID, UUID.randomUUID())
+                        .header("Accept-Language", languageHeader)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isNotFound())
+                .andReturn();
+
+        final var errorMessage = mvcResult.getResponse().getContentAsString(StandardCharsets.UTF_8);
+        assertEquals("User not found", new JSONObject(errorMessage).get("error"));
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"ua", "ua-UA"})
+    @WithMockUser("test")
+    void testUkrainianLocalizationWithAccept_LanguageHeaderIsUA(String locale) throws Exception {
+        MvcResult mvcResult = mockMvc.perform(delete(ADV_ID, UUID.randomUUID())
+                        .header("Accept-Language", locale)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isNotFound())
+                .andReturn();
+
+        final var errorMessage = mvcResult.getResponse().getContentAsString(StandardCharsets.UTF_8);
+        assertEquals("Користувача не знайдено", new JSONObject(errorMessage).get("error"));
     }
 }
