+++ src/test/java/space/obminyashka/items_exchange/controller/UserControllerTest.java
@@ -7,20 +7,15 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.security.core.Authentication;
-import space.obminyashka.items_exchange.model.User;
 import space.obminyashka.items_exchange.service.ImageService;
 import space.obminyashka.items_exchange.service.UserService;
-import space.obminyashka.items_exchange.util.ResponseMessagesHandler;
 
 import java.io.IOException;
 import java.util.Map;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.*;
-import static space.obminyashka.items_exchange.util.MessageSourceUtil.getMessageSource;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class UserControllerTest {
@@ -54,9 +49,8 @@ class UserControllerTest {
 
     @Test
     void updateUserAvatar_WhenReceivedBMPImage_ShouldThrowIOException() {
-        when(userService.findByUsernameOrEmail(anyString())).thenReturn(Optional.of(mock(User.class)));
         when(imageService.scale(bmp)).thenAnswer((t) -> {
-            throw new IOException(getMessageSource(ResponseMessagesHandler.ValidationMessage.INCORRECT_IMAGE_FORMAT));
+            throw new IOException("Incorrect image format");
         });
 
         assertThrows(IOException.class, () -> userController.updateUserAvatar(bmp, authentication));
